Remove pybind11 dependency from cpp library
